[chores] Move logic to utils.py
