+++ requirements-test.txt
@@ -1,1 +1,2 @@
 nose
+coverage
