+++ tests/snmp/airos.py
@@ -5,7 +5,7 @@ from pysnmp.entity.rfc3413.oneliner import cmdgen
 from pysnmp.smi.error import NoSuchObjectError
 
 from ..settings import settings
-from ..static import MockOutputMixin, SpyMock
+from ..utils import MockOutputMixin, SpyMock
 
 
 __all__ = ['TestSNMPAirOS']
@@ -20,7 +20,7 @@ class TestSNMPAirOS(unittest.TestCase, MockOutputMixin):
         self.device = AirOS(self.host, self.community, port=self.port)
 
         # mock calls being made to devices
-        self.oid_mock_data = self._load_mock_json('/test-airos-snmp.json')
+        self.oid_mock_data = self._load_mock_json('/static/test-airos-snmp.json')
         self.nextcmd_patcher = SpyMock._patch(
             target=cmdgen.CommandGenerator,
             attribute='nextCmd',
